Better error handling, needs admin, few more scripts
-Added error handling around invoking Powershell scripts. Any exception
will be caught and the message will be dumped to RichTextBox, and the
next script will begin.
-Better printing to RichTextBox of script output.
-Added checking for errors from script level ("Write-Error
"error_text""). If any are found, they'll be printed to RichTextBox.
-Fixed bug with checking for multiple scripts running at the same time.
-Added scripts for resetting TCP/IP stack, resetting Winsock stack, and
disable/enable WiFi adaptor. Still needs a bit of error checking in
[dis/en]able script, checking if adaptor does not exist.
-Changed app.manifest to require running this application as admin. Many
of the fixes require administrator priveleges, I see no other way around
this.