Strict scripts, formatting, CCleaner script
-Forced all scripts to use StrictMode with latest version to be sure all
variables are declared--no time spent debugging only to find a typo is a
good thing.
-Minor formatting tweaks with logging in RichTextBox.
-Added script for CCleaner. This is functional in debug builds, if
CCleaner is installed. Minor changes will be made for release builds
before it's functional. However--it is not robust! If path/filename for
CCleaner change, this script will break until the globals in it are
changed! This is not sustainable, it will need to be made more robust
before release.