RKill script, formatting
-Yet another minor formatting tweak to log window output.
-Added script for RKill! This will copy "RKill.exe" to the local
machine, rename as "rkill.com", run, and delete local copy. The reason
for copying--the flash drives this utility will be running from is
write-protected. Versions of RKill in the past will refuse to run if
they cannot create their log file, and they generally only attempt to
create on the same drive RKill is run from. This will safeguard against
this bug if RKill regresses in the future.