CheckWRTLogDirs.ps1
-Added script CheckWRTLogDirs.ps1. This will check 2 specific
directories that are prone to fill with a massive amount of log files,
filling the hard drive. If more than 1000MB of log files exist, it will
delete all of them. Script still has an issue--standard Write-Output
function call is not outputting to RichTextBox as expected. Instead,
trying to work out how to print to RichTextBox directly, calling the
Logger.Log function. Still not working as expected, since the text
prints out bolded no matter what.